requirements.txt
flask==3.0.2
flask-socketio==5.3.6
flask-cors==4.0.0
pandas==2.2.0
numpy==1.26.4
plotly==5.18.0
python-dotenv==1.0.1
openai==1.12.0
langchain==0.1.9
langchain-community==0.0.24
langchain-openai==0.0.7
faiss-cpu==1.7.4
werkzeug==3.0.1 